UPDATE-02:
Add an attribute named abilities or instincts, which should be a list type, as an animal can have multiple instincts.

Add a __str__() method for a nicely formatted display of the pet's information.
Set default values for the attributes in the __init__ method if they are not provided.

Create an example where hunger and energy are not given, and the default values are used.